feat: backend-lint-size-004 - Enforce backend size limits in the normal lint and CI workflow
