+++ docs/internal/processes/development-guide-relevant-files.md
@@ -274,6 +274,7 @@ This inventory records the checked-in files and directories that the maintainer 
 - Keep generated replay payload merge regressions in `pkg/replay/event_artifact_test.go` when the behavior under change is deterministic worker or workstation replacement and append ordering; seed `factoryapi.Factory` payloads directly there instead of rebuilding the same coverage through loaded-config filesystem fixtures.
 - When `pkg/replay` loaded-config or event-stream hydration tests grow toward the backend size gate, move generated-factory fixture writers and assertion bundles into sibling same-package helper owners such as `effective_config_generated_test.go` instead of expanding the original behavior-focused test file.
 - When `pkg/listeners/filewatcher_test.go` grows toward the backend size gate, split same-package watcher coverage into sibling files by watched-input concern such as batch normalization, multi-channel directory behavior, and filename-derivation rules while keeping the shared mock factory and submission polling helpers package-local.
+- When helper-heavy runtime files such as `pkg/factory/event_history.go` cross the backend size gate without changing the exported seam, move generated-conversion and projection helpers into sibling same-package owners such as `event_history_generated.go` so the runtime contract stays stable while the maintained file boundary becomes reviewable by concern.
 - Keep safe-diagnostics rehydration ownership in `pkg/interfaces/safe_diagnostics.go`; replay, projection, or worker-facing runtime code should call `interfaces.WorkDiagnosticsFromSafeWorkDiagnostics(...)` instead of rebuilding `WorkDiagnostics` from `SafeWorkDiagnostics` package-locally.
 - When generated safe-diagnostics `StringMap` values need canonical `map[string]string` copies, route the rehydration through `cloneStringMap` inside `pkg/interfaces/safe_diagnostics.go` and normalize zero-length results back to `nil` where that boundary expects nil-for-empty semantics instead of keeping handwritten map-copy loops beside the conversion call sites.
 - Keep worker-facing `interfaces.WorkDiagnostics` deep-clone ownership in `pkg/interfaces`; runtime and replay packages should call the canonical clone helper there instead of maintaining package-local clone trees for nested diagnostics maps, slices, or panic payloads.
