+++ pkg/interfaces/public_factory_enums_test.go
@@ -117,107 +117,7 @@ func TestGeneratedPublicFactoryEnumsPreserveUnknownValues(t *testing.T) {
 }
 
 func TestGeneratedPublicFactoryEnumPtrs(t *testing.T) {
-	tests := []struct {
-		name          string
-		supported     string
-		wantSupported string
-		unknown       string
-		wantUnknown   string
-		ptr           func(string) *string
-	}{
-		{
-			name:          "worker type",
-			supported:     "  MODEL_WORKER  ",
-			wantSupported: "MODEL_WORKER",
-			unknown:       "  CUSTOM_WORKER  ",
-			wantUnknown:   "CUSTOM_WORKER",
-			ptr: func(value string) *string {
-				converted := GeneratedPublicFactoryWorkerTypePtr(value)
-				if converted == nil {
-					return nil
-				}
-				out := string(*converted)
-				return &out
-			},
-		},
-		{
-			name:          "worker model provider",
-			supported:     "  openai  ",
-			wantSupported: "CODEX",
-			unknown:       "  mystery-provider  ",
-			wantUnknown:   "mystery-provider",
-			ptr: func(value string) *string {
-				converted := GeneratedPublicFactoryWorkerModelProviderPtr(value)
-				if converted == nil {
-					return nil
-				}
-				out := string(*converted)
-				return &out
-			},
-		},
-		{
-			name:          "worker provider",
-			supported:     "  local-claude  ",
-			wantSupported: "SCRIPT_WRAP",
-			unknown:       "  custom-executor  ",
-			wantUnknown:   "custom-executor",
-			ptr: func(value string) *string {
-				converted := GeneratedPublicFactoryWorkerProviderPtr(value)
-				if converted == nil {
-					return nil
-				}
-				out := string(*converted)
-				return &out
-			},
-		},
-		{
-			name:          "workstation type",
-			supported:     "  LOGICAL_MOVE  ",
-			wantSupported: "LOGICAL_MOVE",
-			unknown:       "  CUSTOM_WORKSTATION  ",
-			wantUnknown:   "CUSTOM_WORKSTATION",
-			ptr: func(value string) *string {
-				converted := GeneratedPublicFactoryWorkstationTypePtr(value)
-				if converted == nil {
-					return nil
-				}
-				out := string(*converted)
-				return &out
-			},
-		},
-		{
-			name:          "runner id",
-			supported:     "  GEMINI  ",
-			wantSupported: "gemini",
-			unknown:       "  custom-runner  ",
-			wantUnknown:   "custom-runner",
-			ptr: func(value string) *string {
-				converted := GeneratedPublicFactoryRunnerIDPtr(value)
-				if converted == nil {
-					return nil
-				}
-				out := string(*converted)
-				return &out
-			},
-		},
-		{
-			name:          "runner selection source",
-			supported:     "  workstation  ",
-			wantSupported: "workstation",
-			unknown:       "  custom-source  ",
-			wantUnknown:   "custom-source",
-			ptr: func(value string) *string {
-				converted := GeneratedPublicFactoryRunnerSelectionSourcePtr(value)
-				if converted == nil {
-					return nil
-				}
-				out := string(*converted)
-				return &out
-			},
-		},
-	}
-
-	for _, tt := range tests {
+	for _, tt := range generatedPublicFactoryEnumPtrCases() {
 		t.Run(tt.name, func(t *testing.T) {
 			if got := tt.ptr("   "); got != nil {
 				t.Fatalf("expected nil pointer for whitespace-only input")
@@ -238,4 +138,98 @@ func TestGeneratedPublicFactoryEnumPtrs(t *testing.T) {
 			}
 		})
 	}
+}
+
+type generatedPublicFactoryEnumPtrCase struct {
+	name          string
+	supported     string
+	wantSupported string
+	unknown       string
+	wantUnknown   string
+	ptr           func(string) *string
+}
+
+func generatedPublicFactoryEnumPtrCases() []generatedPublicFactoryEnumPtrCase {
+	return []generatedPublicFactoryEnumPtrCase{
+		{
+			name:          "worker type",
+			supported:     "  MODEL_WORKER  ",
+			wantSupported: "MODEL_WORKER",
+			unknown:       "  CUSTOM_WORKER  ",
+			wantUnknown:   "CUSTOM_WORKER",
+			ptr:           generatedPublicFactoryWorkerTypeStringPtr,
+		},
+		{
+			name:          "worker model provider",
+			supported:     "  openai  ",
+			wantSupported: "CODEX",
+			unknown:       "  mystery-provider  ",
+			wantUnknown:   "mystery-provider",
+			ptr:           generatedPublicFactoryWorkerModelProviderStringPtr,
+		},
+		{
+			name:          "worker provider",
+			supported:     "  local-claude  ",
+			wantSupported: "SCRIPT_WRAP",
+			unknown:       "  custom-executor  ",
+			wantUnknown:   "custom-executor",
+			ptr:           generatedPublicFactoryWorkerProviderStringPtr,
+		},
+		{
+			name:          "workstation type",
+			supported:     "  LOGICAL_MOVE  ",
+			wantSupported: "LOGICAL_MOVE",
+			unknown:       "  CUSTOM_WORKSTATION  ",
+			wantUnknown:   "CUSTOM_WORKSTATION",
+			ptr:           generatedPublicFactoryWorkstationTypeStringPtr,
+		},
+		{
+			name:          "runner id",
+			supported:     "  GEMINI  ",
+			wantSupported: "gemini",
+			unknown:       "  custom-runner  ",
+			wantUnknown:   "custom-runner",
+			ptr:           generatedPublicFactoryRunnerIDStringPtr,
+		},
+		{
+			name:          "runner selection source",
+			supported:     "  workstation  ",
+			wantSupported: "workstation",
+			unknown:       "  custom-source  ",
+			wantUnknown:   "custom-source",
+			ptr:           generatedPublicFactoryRunnerSelectionSourceStringPtr,
+		},
+	}
+}
+
+func generatedPublicFactoryWorkerTypeStringPtr(value string) *string {
+	return generatedPublicFactoryStringPtr(GeneratedPublicFactoryWorkerTypePtr(value))
+}
+
+func generatedPublicFactoryWorkerModelProviderStringPtr(value string) *string {
+	return generatedPublicFactoryStringPtr(GeneratedPublicFactoryWorkerModelProviderPtr(value))
+}
+
+func generatedPublicFactoryWorkerProviderStringPtr(value string) *string {
+	return generatedPublicFactoryStringPtr(GeneratedPublicFactoryWorkerProviderPtr(value))
+}
+
+func generatedPublicFactoryWorkstationTypeStringPtr(value string) *string {
+	return generatedPublicFactoryStringPtr(GeneratedPublicFactoryWorkstationTypePtr(value))
+}
+
+func generatedPublicFactoryRunnerIDStringPtr(value string) *string {
+	return generatedPublicFactoryStringPtr(GeneratedPublicFactoryRunnerIDPtr(value))
+}
+
+func generatedPublicFactoryRunnerSelectionSourceStringPtr(value string) *string {
+	return generatedPublicFactoryStringPtr(GeneratedPublicFactoryRunnerSelectionSourcePtr(value))
+}
+
+func generatedPublicFactoryStringPtr[T ~string](value *T) *string {
+	if value == nil {
+		return nil
+	}
+	out := string(*value)
+	return &out
 }
